rename folder datastore -> remote_client
